#163 TopNavigation rework
