+++ src/components/MobileNavigation/MobileNavigation.module.scss
@@ -1,0 +1,33 @@
+.wrapper {
+  display: none;
+
+  @media only screen and (width <= 768px) {
+    display: block;
+  }
+
+  &__menu-stripe {
+    padding: var(--space-20-40) var(--space-30-40);
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+  }
+}
+
+.menu-button {
+  display: block;
+  border: none;
+  background: none;
+
+  &:hover {
+    cursor: pointer;
+  }
+}
+
+.links {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-10-15);
+  list-style: none;
+  padding-left: var(--space-30-40);
+  width: max-content;
+}
