--- src/components/TopNavigation/TopNavigation.module.scss
@@ -1,49 +1,0 @@
-.wrapper {
-  padding: var(--space-20-40) var(--space-40-60-wide);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  @media only screen and (width <= 1480px) {
-    padding: var(--space-20-40) var(--space-30-40);
-  }
-}
-
-.links {
-  &--desktop {
-    display: flex;
-    list-style: none;
-    gap: var(--space-40-50);
-
-    @media only screen and (width <= 768px) {
-      display: none;
-    }
-  }
-
-  &--mobile {
-    display: none;
-
-    @media only screen and (width <= 768px) {
-      display: flex;
-      flex-direction: column;
-      gap: var(--space-10-15);
-      list-style: none;
-      padding-left: var(--space-30-40);
-      width: max-content;
-    }
-  }
-}
-
-.menu-button {
-  display: none;
-
-  &:hover {
-    cursor: pointer;
-  }
-
-  @media only screen and (width <= 768px) {
-    display: block;
-    border: none;
-    background: none;
-  }
-}
